Moved unittest to the corresponding namespace.
